+++ packages/stream_chat_persistence/lib/src/dao/pinned_message_reaction_dao.g.dart
@@ -4,7 +4,6 @@ part of 'pinned_message_reaction_dao.dart';
 
 // ignore_for_file: type=lint
 mixin _$PinnedMessageReactionDaoMixin on DatabaseAccessor<DriftChatDatabase> {
-  $PinnedMessagesTable get pinnedMessages => attachedDatabase.pinnedMessages;
   $PinnedMessageReactionsTable get pinnedMessageReactions => attachedDatabase.pinnedMessageReactions;
   $UsersTable get users => attachedDatabase.users;
   PinnedMessageReactionDaoManager get managers => PinnedMessageReactionDaoManager(this);
@@ -13,10 +12,6 @@ mixin _$PinnedMessageReactionDaoMixin on DatabaseAccessor<DriftChatDatabase> {
 class PinnedMessageReactionDaoManager {
   final _$PinnedMessageReactionDaoMixin _db;
   PinnedMessageReactionDaoManager(this._db);
-  $$PinnedMessagesTableTableManager get pinnedMessages => $$PinnedMessagesTableTableManager(
-    _db.attachedDatabase,
-    _db.pinnedMessages,
-  );
   $$PinnedMessageReactionsTableTableManager get pinnedMessageReactions => $$PinnedMessageReactionsTableTableManager(
     _db.attachedDatabase,
     _db.pinnedMessageReactions,