+++ packages/stream_chat_persistence/lib/src/dao/poll_vote_dao.g.dart
@@ -4,7 +4,6 @@ part of 'poll_vote_dao.dart';
 
 // ignore_for_file: type=lint
 mixin _$PollVoteDaoMixin on DatabaseAccessor<DriftChatDatabase> {
-  $PollsTable get polls => attachedDatabase.polls;
   $PollVotesTable get pollVotes => attachedDatabase.pollVotes;
   $UsersTable get users => attachedDatabase.users;
   PollVoteDaoManager get managers => PollVoteDaoManager(this);
@@ -13,7 +12,6 @@ mixin _$PollVoteDaoMixin on DatabaseAccessor<DriftChatDatabase> {
 class PollVoteDaoManager {
   final _$PollVoteDaoMixin _db;
   PollVoteDaoManager(this._db);
-  $$PollsTableTableManager get polls => $$PollsTableTableManager(_db.attachedDatabase, _db.polls);
   $$PollVotesTableTableManager get pollVotes => $$PollVotesTableTableManager(_db.attachedDatabase, _db.pollVotes);
   $$UsersTableTableManager get users => $$UsersTableTableManager(_db.attachedDatabase, _db.users);
 }