+++ packages/stream_chat_persistence/lib/src/db/drift_chat_database.dart
@@ -24,6 +24,7 @@ part 'drift_chat_database.g.dart';
     Members,
     Reads,
     ChannelQueries,
+    ChannelQueriesMetadata,
     ConnectionEvents,
   ],
   daos: [